ci: drop ccache usage
It doesn't seem to work for all runs as expected, sometimes the install
step takes almost as much time as running without ccache, and cache
handling is poor (outdated caches aren't removed by default).

Overall it doesn't appear to be a net benefit for the current project's
size and scope at this time.