don't raise exceptions from validation function by default
Exceptions can be useful when the reason why the parsing failure is
wanted, but in general the functions are to determine the boolean
validity of strings for a given type.

Now the `raised` parameter has to be enabled if an exception should be
raised on failure, otherwise False is returned by default.