dep: cache more Dep attributes
This adds ~20-30 ns to instance creation but speeds up multiple attr
access by ~3x for string-based attrs and ~20x for more generic objects.

Whether this is worth it or not depends on how much each attr is used,
but doing this for more attrs feels more consistent as the same fields
are cached in the underlying rust implementation.